Adding TextActivity.java and activity_text.xml as UI for chatbot to free up the MainActivity.java and activity_main.xml
