+++ app/src/main/res/layout/activity_chatbot.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".TextActivity">
 
     <ListView
         android:id="@+id/listView"
