+++ app/src/main/res/raw/bitbotContent.csv
